+++ src/integration/groovy/pl/allegro/tech/leaders/hackathon/base/IntegrationSpec.groovy
@@ -1,4 +1,4 @@
-package integration.base
+package pl.allegro.tech.leaders.hackathon.base
 
 import groovy.transform.CompileStatic
 import org.springframework.beans.factory.annotation.Autowired
